+++ db/config/database.json
@@ -1,7 +1,7 @@
 {
   "development": {
-    "username": "kisinoverskaya",
-    "password": "123456",
+    "username": "Vadim",
+    "password": "12345",
     "database": "login",
     "host": "127.0.0.1",
     "dialect": "postgres"
